backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Fruitful Lab with SQLAlchemy and Postgres"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.123.5",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
]

[build-system]
requires = ["hatchling>=1.18"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Tell hatch which files to include in the wheel.
# Our package is the current directory (backend/).
packages = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
